Cardano.Address

Address

data Address

An Address type representing Cardano addresses. Internals are irrevelant to the user.

Since: 1.0.0

class PaymentAddress key where

Encoding of addresses for certain key types and backend targets.

Since: 1.0.0

Methods

paymentAddress :: NetworkDiscriminant -> key AddressK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Since: 1.0.0

Conversion From / To Text

base58 :: Address -> Text

Encode an Address to a base58 Text.

Since: 1.0.0

fromBase58 :: Text -> Maybe Address

Decode a base58-encoded Text into an Address

Since: 1.0.0

bech32 :: Address -> Text

Encode an Address to bech32 Text, using addr as a human readable prefix.

Since: 1.0.0

fromBech32 :: Text -> Maybe Address

Decode a bech32-encoded Text into an Address

Since: 1.0.0

Network Discrimination

data NetworkDiscriminant

Describe required network discrimination. See also the available smart-constructors if you are not sure about what to do with this:

mainnetDiscriminant

stagingDiscriminant

testnetDiscriminant

Since: 1.0.0

Constructors

RequiresNoMagic

RequiresMagic ProtocolMagic

mainnetDiscriminant :: NetworkDiscriminant

Required network discrimination settings for mainnet

Since: 1.0.0

stagingDiscriminant :: NetworkDiscriminant

Required network discrimination settings for staging

Since: 1.0.0

testnetDiscriminant :: NetworkDiscriminant

Required network discrimination settings for testnet

Since: 1.0.0

Protocol Magic

newtype ProtocolMagic

Magic constant associated with a given network. This is mainly used in two places:

In Address payloads, to discriminate addresses between networks.

At the network-level, when doing handshake with nodes.

Since: 1.0.0

Constructors

ProtocolMagic Word32

mainnetMagic :: ProtocolMagic

Hard-coded ProtocolMagic for Cardano MainNet

Since: 1.0.0

stagingMagic :: ProtocolMagic

Hard-coded ProtocolMagic for Cardano Staging

Since: 1.0.0

testnetMagic :: ProtocolMagic

Hard-coded ProtocolMagic for Cardano standard TestNet

Since: 1.0.0