Cardano.Address.Compat

Description

A compatibility function for the GHCJS build.

Documentation

ghcjsBuildSupport :: IO ()

This function must be used somewhere, so that external Javascript files are correctly linked in the GHCJS build.

For non-GHCJS, it has no effect.