Cardano.Address.Derivation

Description

Overview

These abstractions allow generating root private key, also called Master Key and then basing on it enable address derivation

Key Derivation

Types

data Index (derivationType :: DerivationType) (depth :: Depth)

A derivation index, with phantom-types to disambiguate derivation type.

let accountIx = Index 'Hardened 'AccountK
let addressIx = Index 'Soft 'PaymentK

Since: 1.0.0

indexToWord32 :: Index derivationType depth -> Word32

Get the index as a Word32 @since 3.3.0

indexFromWord32 :: forall ix derivationType depth. (ix ~ Index derivationType depth, Bounded ix) => Word32 -> Maybe ix

Construct derivation path indices from raw Word32 values.

wholeDomainIndex :: Word32 -> Index 'WholeDomain depth

Constructs a full domain Index. This can't fail, unlike fromWord32.

Since: 3.3.0

coerceWholeDomainIndex :: Index ty depth0 -> Index 'WholeDomain depth1

Upcasts an Index to one with the full Word32 domain.

Since: 3.3.0

nextIndex :: forall ix derivationType depth. (ix ~ Index derivationType depth, Bounded ix) => ix -> Maybe ix

Increment an index, if possible.

Since: 3.3.0

data Depth

Key Depth in the derivation path, according to BIP-0039 and BIP-0044.

root | purpose' | cointype' | account' | role | address
0th      1st         2nd        3rd       4th     5th

We do not manipulate purpose, cointype and change paths directly, so there are no constructors for these.

Since: 1.0.0

Constructors

RootK

AccountK

PaymentK

DelegationK

DRepK

CCColdK

CCHotK

ScriptK

PolicyK

data DerivationType

Type of derivation that should be used with the given indexes.

In theory, we should only consider two derivation types: soft and hard.

However, historically, addresses in Cardano used to be generated across both the soft and the hard domain. We therefore introduce a WholeDomain derivation type that is the exact union of Hardened and Soft.

Since: 1.0.0

Constructors

Hardened

Soft

WholeDomain

Abstractions

class GenMasterKey (key :: Depth -> Type -> Type) where

Abstract interface for constructing a Master Key.

Since: 1.0.0

Associated Types

type SecondFactor key :: Type

Methods

genMasterKeyFromMnemonic :: SomeMnemonic -> SecondFactor key -> key 'RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 1.0.0

genMasterKeyFromXPrv :: XPrv -> key 'RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 1.0.0

class HardDerivation (key :: Depth -> Type -> Type) where

An interface for doing hard derivations from the root private key, Master Key

Since: 1.0.0

Associated Types

type AccountIndexDerivationType key :: DerivationType

type AddressIndexDerivationType key :: DerivationType

type WithRole key :: Type

Methods

deriveAccountPrivateKey :: key 'RootK XPrv -> Index (AccountIndexDerivationType key) 'AccountK -> key 'AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

Since: 1.0.0

deriveAddressPrivateKey :: key 'AccountK XPrv -> WithRole key -> Index (AddressIndexDerivationType key) 'PaymentK -> key 'PaymentK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

Since: 1.0.0

class HardDerivation key => SoftDerivation (key :: Depth -> Type -> Type) where

An interface for doing soft derivations from an account public key

Methods

deriveAddressPublicKey :: key 'AccountK XPub -> WithRole key -> Index 'Soft 'PaymentK -> key 'PaymentK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Since: 1.0.0

Low-Level Cryptography Primitives

XPrv

xprvFromBytes :: ByteString -> Maybe XPrv

Construct an XPrv from raw ByteString (96 bytes).

Since: 1.0.0

xprvToBytes :: XPrv -> ByteString

Convert an XPrv to a raw ByteString (96 bytes).

Since: 1.0.0

xprvPrivateKey :: XPrv -> ByteString

Extract the private key from an XPrv as a raw ByteString (64 bytes).

Since: 2.0.0

xprvChainCode :: XPrv -> ByteString

Extract the chain code from an XPrv as a raw ByteString (32 bytes).

Since: 2.0.0

toXPub :: XPrv -> XPub

Derive the XPub associated with an XPrv.

Since: 1.0.0

XPub

xpubFromBytes :: ByteString -> Maybe XPub

An opaque type representing an extended private key.

Properties:

Roundtripping

forall xprv. xprvFromBytes (xprvToBytes xprv) == Just xprv

Chain Code Invariance

forall xprv. xprvChainCode xprv == xpubChainCode (toXPub xprv)

Public Key Signature

forall xprv msg. verify (toXPub xprv) msg (sign xprv msg) == True

NOTE: XPrv, XPub, XSignature are re-exported from Cardano.Address.Crypto

Construct an XPub from raw ByteString (64 bytes).

Since: 1.0.0

xpubToBytes :: XPub -> ByteString

Convert an XPub to a raw ByteString (64 bytes).

Since: 1.0.0

xpubPublicKey :: XPub -> ByteString

Extract the public key from an XPub as a raw ByteString (32 bytes).

Since: 2.0.0

xpubChainCode :: XPub -> ByteString

Extract the chain code from an XPub as a raw ByteString (32 bytes).

Since: 2.0.0

Pub

data Pub

An opaque type representing a non-extended public key.

Properties:

Roundtripping

forall pub. pubFromBytes (pubToBytes pub) == Just pub

Since: 3.12.0

pubFromBytes :: ByteString -> Maybe Pub

Construct a Pub from raw ByteString (32 bytes).

Since: 3.12.0

pubToBytes :: Pub -> ByteString

Convert an Pub to a raw ByteString (32 bytes).

Since: 3.12.0

xpubToPub :: XPub -> Pub

Extract the public key from an XPub as a Pub (32 bytes).

Since: 3.12.0

XSignature

sign :: ByteArrayAccess msg => XPrv -> msg -> XSignature

Produce a signature of the given msg from an XPrv.

Since: 1.0.0

verify :: ByteArrayAccess msg => XPub -> msg -> XSignature -> Bool

Verify the XSignature of a msg with the XPub associated with the XPrv used for signing.

Since: 1.0.0