Cardano.Address.Script

Description

Script

data Script (elem :: Type)

A Script type represents multi signature script. The script embodies conditions that need to be satisfied to make it valid.

Since: 3.0.0

Constructors

RequireSignatureOf !elem

RequireAllOf ![Script elem]

RequireAnyOf ![Script elem]

RequireSomeOf Word8 ![Script elem]

ActiveFromSlot Natural

ActiveUntilSlot Natural

serializeScript :: Script KeyHash -> ByteString

This function realizes what cardano-node's `Api.serialiseToCBOR script` realizes This is basically doing the symbolically following: toCBOR [0,multisigScript]

Since: 3.0.0

foldScript :: (a -> b -> b) -> b -> Script a -> b

Script folding

Since: 3.2.0

Script template

data ScriptTemplate

Represents the script template that show the structure of the script and determines the expected place of verification keys corresponding to given cosigners.

Since: 3.2.0

Constructors

ScriptTemplate

Fields

cosigners :: Map Cosigner XPub

template :: Script Cosigner

newtype Cosigner

Represents the cosigner of the script, ie., party that co-shares the script.

Since: 3.2.0

Constructors

Cosigner Word8

Validation

data ValidationLevel

Validation level. Required level does basic check that will make sure the script is accepted in ledger. Recommended level collects a number of checks that will warn about dangerous, unwise and redundant things present in the script.

Since: 3.2.0

Constructors

RequiredValidation

RecommendedValidation

data ErrValidateScript

Possible validation errors when validating a script

Since: 3.0.0

Constructors

LedgerIncompatible

WrongKeyHash

NotUniformKeyType

Malformed

NotRecommended ErrRecommendedValidateScript

data ErrRecommendedValidateScript

Possible recommended validation errors when validating a script

Since: 3.2.0

Constructors

EmptyList

ListTooSmall

MZero

DuplicateSignatures

RedundantTimelocks

TimelockTrap

data ErrValidateScriptTemplate

Possible validation errors when validating a script template

Since: 3.2.0

Constructors

WrongScript ErrValidateScript

DuplicateXPubs

UnknownCosigner

MissingCosignerXPub

NoCosignerInScript

NoCosignerXPub

validateScript :: ValidationLevel -> Script KeyHash -> Either ErrValidateScript ()

Validate a Script, semantically

Since: 3.0.0

validateScriptTemplate :: ValidationLevel -> ScriptTemplate -> Either ErrValidateScriptTemplate ()

Validate a ScriptTemplate, semantically

Since: 3.2.0

validateScriptOfTemplate :: ValidationLevel -> Script Cosigner -> Either ErrValidateScript ()

Validate a script in ScriptTemplate

Since: 3.5.0

prettyErrValidateScript :: ErrValidateScript -> String

Pretty-print a script validation error.

Since: 3.0.0

prettyErrValidateScriptTemplate :: ErrValidateScriptTemplate -> String

Pretty-print a script template validation error.

Since: 3.2.0

Hashing

newtype ScriptHash

A ScriptHash type represents script hash. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

ScriptHash

Fields

unScriptHash :: ByteString

toScriptHash :: Script KeyHash -> ScriptHash

Computes the hash of a given script, by first serializing it to CBOR.

Since: 3.0.0

scriptHashFromBytes :: ByteString -> Maybe ScriptHash

Construct an ScriptHash from raw ByteString (28 bytes).

Since: 3.0.0

scriptHashToText :: ScriptHash -> KeyRole -> Maybe GovernanceType -> Text

Encode a ScriptHash to bech32 Text or hex if key role is unknown. In the case of governance role, if one wants to include additional byte as specified in CIP-0129 unless the function is called with CIP0105.

One byte is prepended to script hash only in governance context. The rules how to contruct it are summarized below

  drep       0010....
  hot        0000....    key type
  cold       0001....

  scripthash ....0011    credential type

This is on top of X_script, where X={drep, cc_hot, cc_hot}, which lacks the additional byte. In scriptHashFromText we additionally support reading legacy X which also lacks the additional byte, and has the same payload as as the corresponding X_script.

Since: 4.0.0

scriptHashFromText :: Text -> Either ErrScriptHashFromText ScriptHash

Construct a ScriptHash from Text. It should be

Bech32 encoded text with one of following hrp:

script

drep

cc_cold

cc_hot

drep_script

cc_cold_script

cc_hot_script

If if hex is encountered it is converted in rawly fashion

Since: 4.0.0

prettyErrScriptHashFromText :: ErrScriptHashFromText -> String

Possible errors when deserializing a script hash from text.

Since: 4.0.0