Cardano.Address.Style.Byron

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Byron master keys from mnemonic sentences

HardDerivation: for hierarchical derivation of parent to child keys

PaymentAddress: for constructing addresses from a public key

We call Byron addresses the old address type used by Daedalus in the early days of Cardano. Using this type of addresses and underlying key scheme is now considered deprecated because of some security implications.

The internals of the Byron does not matter for the reader, but basically contains what is necessary to perform key derivation and generate addresses from a Byron type.

Deprecation Notice

Unless you have good reason to do so (like writing backward-compatible code with an existing piece), any new implementation should use the Icarus style for key and addresses.

Byron

data Byron (depth :: Depth) key

Deprecated: see Icarus

Material for deriving HD random scheme keys, which can be used for making addresses.

Since: 1.0.0

type family DerivationPath (depth :: Depth) :: Type where ...

Deprecated: see Icarus

The hierarchical derivation indices for a given level/depth.

Since: 1.0.0

Equations

DerivationPath 'RootK = ()

DerivationPath 'AccountK = Index 'WholeDomain 'AccountK

DerivationPath 'PaymentK = (Index 'WholeDomain 'AccountK, Index 'WholeDomain 'PaymentK)

$sel:payloadPassphrase:Byron :: Byron depth key -> ScrubbedBytes

Deprecated: see Icarus

Used for encryption of the derivation path payload within an address.

Since: 1.0.0

$sel:derivationPath:Byron :: Byron depth key -> DerivationPath depth

Deprecated: see Icarus

The address derivation indices for the level of this key.

Since: 1.0.0

$sel:getKey:Byron :: Byron depth key -> key

Deprecated: see Icarus

The raw private or public key.

Since: 1.0.0

Key Derivation

genMasterKeyFromXPrv :: XPrv -> Byron 'RootK XPrv

Deprecated: see Icarus

Generate a root key from a corresponding root XPrv

Since: 1.0.0

genMasterKeyFromMnemonic

Arguments

:: SomeMnemonic
Some valid mnemonic sentence.

-> Byron 'RootK XPrv

Deprecated: see Icarus

Generate a root key from a corresponding mnemonic.

Since: 1.0.0

deriveAccountPrivateKey :: Byron 'RootK XPrv -> Index 'WholeDomain 'AccountK -> Byron 'AccountK XPrv

Deprecated: see Icarus

Derives an account private key from the given root private key.

Since: 1.0.0

deriveAddressPrivateKey :: Byron 'AccountK XPrv -> Index 'WholeDomain 'PaymentK -> Byron 'PaymentK XPrv

Deprecated: see Icarus

Derives an address private key from the given account private key.

Since: 1.0.0

Addresses

Generating a PaymentAddress

import Cardano.Address ( base58 )
import Cardano.Address.Derivation ( toXPub(..) )

base58 $ paymentAddress byronMainnet (toXPub <$> addrK)
"DdzFFzCqrhsq3KjLtT51mESbZ4RepiHPzLqEhamexVFTJpGbCXmh7qSxnHvaL88QmtVTD1E1sjx8Z1ZNDhYmcBV38ZjDST9kYVxSkhcw"

data AddressInfo

The result of eitherInspectAddress for Byron addresses.

Since: 3.4.0

Constructors

AddressInfo

Fields

infoAddressRoot :: !ByteString

infoPayload :: !PayloadInfo

infoNetworkTag :: !(Maybe NetworkTag)

eitherInspectAddress :: Maybe XPub -> Address -> Either ErrInspectAddress AddressInfo

Determines whether an Address is a Byron address.

Returns either details about the Address, or ErrInspectAddress if it's not a valid address.

Since: 3.4.0

inspectAddress :: forall m. MonadThrow m => Maybe XPub -> Address -> m Value

Determines whether an Address is a Byron address.

Returns a JSON object with information about the address, or throws ErrInspectAddress if the address isn't a byron address.

Since: 3.0.0

paymentAddress :: NetworkDiscriminant Byron -> Byron 'PaymentK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Since: 1.0.0

data ErrInspectAddress

Possible errors from inspecting a Byron address

Since: 3.0.0

Constructors

MissingExpectedDerivationPath

DeserialiseError DeserialiseFailure

FailedToDecryptPath

prettyErrInspectAddress :: ErrInspectAddress -> String

Pretty-print an ErrInspectAddress

Since: 3.0.0

Network Discrimination

byronMainnet :: NetworkDiscriminant Byron

NetworkDiscriminant for Cardano MainNet & Byron

Since: 2.0.0

byronStaging :: NetworkDiscriminant Byron

NetworkDiscriminant for Cardano Staging & Byron

Since: 2.0.0

byronTestnet :: NetworkDiscriminant Byron

NetworkDiscriminant for Cardano Testnet & Byron

Since: 2.0.0

byronPreprod :: NetworkDiscriminant Byron

NetworkDiscriminant for Cardano Preprod & Byron

Since: 3.13.0

byronPreview :: NetworkDiscriminant Byron

NetworkDiscriminant for Cardano Preview & Byron

Since: 3.13.0

Unsafe

liftXPrv

Arguments

:: XPub
A root public key

-> DerivationPath depth

-> XPrv

-> Byron depth XPrv

Deprecated: see Icarus

Backdoor for generating a new key from a raw XPrv.

Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should genMasterKeyFromXPrv

The first argument is a type-family DerivationPath and its type depends on the depth of the key.

examples:

>>> liftXPrv rootPrv () prv
_ :: Byron RootK XPrv

>>> liftXPrv rootPrv minBound prv
_ :: Byron AccountK XPrv

>>> liftXPrv rootPrv (minBound, minBound) prv
_ :: Byron PaymentK XPrv

Since: 2.0.0

liftXPub

Arguments

:: XPub
A root public key

-> DerivationPath depth

-> XPub

-> Byron depth XPub

Deprecated: see Icarus

Backdoor for generating a new key from a raw XPub.

Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should genMasterKeyFromXPrv

see also liftXPrv

Since: 2.0.0