Cardano.Address.Style.Jormungandr

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Jormungandr master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

PaymentAddress: for constructing payment addresses from a address public key

DelegationAddress: for constructing payment addresses from a address and stake public keys

Disclaimer

Beware that this format of address have been specially crafted for the incentivized testnet (ITN). This will soon be deprecated in favor of Shelley.

New integrations should not use this address style.

Jormungandr

data Jormungandr (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Jormungandr 'RootK XPrv
let accountPubKey  = Jormungandr 'AccountK XPub
let addressPubKey  = Jormungandr 'AddressK XPub

Since: 2.0.0

getKey :: Jormungandr depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 2.0.0

Key Derivation

Generating a root key from SomeMnemonic

:set -XOverloadedStrings
:set -XTypeApplications
:set -XDataKinds
import Cardano.Mnemonic ( mkSomeMnemonic )

let (Right mw) = mkSomeMnemonic @'[15] ["network","empty","cause","mean","expire","private","finger","accident","session","problem","absurd","banner","stage","void","what"]
let sndFactor = mempty -- Or alternatively, a second factor mnemonic transformed to bytes via someMnemonicToBytes
let rootK = genMasterKeyFromMnemonic mw sndFactor :: Jormungandr 'RootK XPrv

Deriving child keys

Let's consider the following 3rd, 4th and 5th derivation paths 0'/0/14

import Cardano.Address.Derivation ( AccountingStyle(..) )

let accIx = toEnum 0x80000000
let acctK = deriveAccountPrivateKey rootK accIx

let addIx = toEnum 0x00000014
let addrK = deriveAddressPrivateKey acctK UTxOExternal addIx

let stakeK = deriveStakingPrivateKey acctK

genMasterKeyFromXPrv :: XPrv -> Jormungandr RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 2.0.0

genMasterKeyFromMnemonic

Arguments

:: SomeMnemonic
Some valid mnemonic sentence.

-> ScrubbedBytes
An optional second-factor passphrase (or mempty)

-> Jormungandr RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 2.0.0

deriveAccountPrivateKey :: Jormungandr RootK XPrv -> Index Hardened AccountK -> Jormungandr AccountK XPrv

Derives an account private key from the given root private key.

Since: 2.0.0

deriveAddressPrivateKey :: Jormungandr AccountK XPrv -> AccountingStyle -> Index Soft AddressK -> Jormungandr AddressK XPrv

Derives an address private key from the given account private key.

Since: 2.0.0

deriveAddressPublicKey :: Jormungandr AccountK XPub -> AccountingStyle -> Index Soft AddressK -> Jormungandr AddressK XPub

Derives an address public key from the given account public key.

Since: 2.0.0

deriveStakingPrivateKey :: Jormungandr AccountK XPrv -> Jormungandr StakingK XPrv

Derive a staking key for a corresponding AccountK. Note that wallet software are by convention only using one staking key per account, and always the first account (with index 0').

Deriving staking keys for something else than the initial account is not recommended and can lead to incompatibility with existing wallet softwares (Daedalus, Yoroi, Adalite...).

Since: 2.0.0

Addresses

Generating a PaymentAddress

import Cardano.Address ( bech32 )
import Cardano.Address.Derivation ( AccountingStyle(..), toXPub(..) )

bech32 $ paymentAddress incentivizedTestnet (toXPub <$> addrK)
"addr1vxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdncxsce5t"

Generating a DelegationAddress

import Cardano.Address ( DelegationAddress (..) )
import Cardano.Address.Derivation ( StakingDerivation (..) )

bech32 $ delegationAddress incentivizedTestnet (toXPub <$> addrK) (toXPub <$> stakeK)
"addr1qxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdn7nudck0fzve4346yytz3wpwv9yhlxt7jwuc7ytwx2vfkyqmkc5xa"

inspectJormungandrAddress :: Address -> Maybe Value

Analyze an Address to know whether it's a Jormungandr address or not.

Returns Nothing if it's not a valid Shelley address, or a ready-to-print string giving details about the Jormungandr.

Since: 2.0.0

paymentAddress :: NetworkDiscriminant Jormungandr -> Jormungandr AddressK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Since: 2.0.0

delegationAddress :: NetworkDiscriminant Jormungandr -> Jormungandr AddressK XPub -> Jormungandr StakingK XPub -> Address

Convert a public key and a staking key to a delegation Address valid for the given network discrimination. Funds sent to this address will be delegated according to the delegation settings attached to the delegation key.

Since: 2.0.0

Network Discrimination

incentivizedTestnet :: NetworkDiscriminant Jormungandr

NetworkDiscriminant for Cardano Incentivized Testnet testnet & Jormungandr

Since: 2.0.0

Unsafe

liftXPrv :: XPrv -> Jormungandr depth XPrv

Unsafe backdoor for constructing an Jormungandr key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0