Cardano.AddressDerivation.Byron

Description

Implementation of address derivation for the random scheme, as implemented by the legacy Cardano wallets.

For full documentation of the key derivation schemes, see the Cardano.Crypto.Wallet module, and the implementation in cardano-crypto.

Types

data Byron (depth :: Depth) key

Material for deriving HD random scheme keys, which can be used for making addresses.

Constructors

Byron

Fields

getKey :: key

The raw private or public key.

derivationPath :: DerivationPath depth

The address derivation indices for the level of this key.

payloadPassphrase :: ScrubbedBytes

Used for encryption of payload containing address derivation path.

Generation

unsafeGenerateKeyFromSeed :: DerivationPath depth -> SomeMnemonic -> ScrubbedBytes -> Byron depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The amount of entropy carried by a BIP-39 12-word mnemonic is 16 bytes.

unsafeMkByronKeyFromMasterKey :: DerivationPath depth -> XPrv -> Byron depth XPrv

mkByronKeyFromMasterKey :: XPrv -> Byron RootK XPrv