Cardano.AddressDerivation.Icarus

Description

Implementation of address derivation for the random scheme, as implemented by the Icarus wallet.

For full documentation of the key derivation schemes, see the Cardano.Crypto.Wallet module, and the implementation in cardano-crypto.

Types

newtype Icarus (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Icarus 'RootK XPrv
let accountPubKey = Icarus 'AccountK XPub
let addressPubKey = Icarus 'AddressK XPub

Constructors

Icarus

Fields

getKey :: key

Generation

unsafeGenerateKeyFromHardwareLedger

Arguments

:: SomeMnemonic
The root mnemonic

-> ScrubbedBytes
Master encryption passphrase

-> Icarus RootK XPrv

Hardware Ledger devices generates keys from mnemonic using a different approach (different from the rest of Cardano).

It is a combination of:

SLIP 0010

BIP 0032

BIP 0039

RFC 8032

What seems to be arbitrary changes from Ledger regarding the calculation of the initial chain code and generation of the root private key.

unsafeGenerateKeyFromSeed

Arguments

:: SomeMnemonic
The root mnemonic

-> ScrubbedBytes
Master encryption passphrase

-> Icarus depth XPrv

Generate a new key from seed. Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should use generateKeyFromSeed.

minSeedLengthBytes :: Int

The minimum seed length for generateKeyFromSeed and unsafeGenerateKeyFromSeed.

Helpers

publicKey :: Icarus depth1 XPrv -> Icarus depth2 XPub