Cardano.Script

Documentation

data Script

A Script type represents multi signature script. The script embodies conditions that need to be satisfied to make it valid.

Since: 3.0.0

Constructors

RequireSignatureOf KeyHash

RequireAllOf [Script]

RequireAnyOf [Script]

RequireMOf Word8 [Script]

newtype KeyHash

A KeyHash type represents verification key hash that participate in building MultisigScript. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

KeyHash ByteString

hashKey :: Shelley key XPub -> KeyHash

Apply Blake2b224 hashing on Shelley XPub. This results in 28-byte hash.

Since: 3.0.0

keyHashFromBytes :: ByteString -> Maybe KeyHash

Construct an KeyHash from raw ByteString (28 bytes).

Since: 3.0.0

toScriptHash :: Script -> ByteString

This function realizes what cardano-node's `Api.serialiseToRawBytes $ Api.scriptHash script` realizes This is basically doing the symbolically the following (using symbols from toCBOR): digest $ (nativeMultisigTag <> toCBOR multisigScript )

validateScript :: Script -> Bool

Validate Script

Since: 3.0.0

Internal

toCBOR :: Script -> ByteString

This function realizes what cardano-node's `Api.serialiseToCBOR script` realizes This is basically doing the symbolically following: toCBOR [0,multisigScript]