Cardano.Address.Derivation

Overview

These abstractions allow generating root private key, also called Master Key and then basing on it enable address derivation

Abstractions

class GenMasterKey (key :: Depth -> * -> *) where

Abstract interface for constructing a Master Key.

Associated Types

type GenMasterKeyFrom key :: *

Methods

genMasterKey :: GenMasterKeyFrom key -> ScrubbedBytes -> key RootK XPrv

Generate a root key from a corresponding seed.

class HardDerivation (key :: Depth -> * -> *) where

An interface for doing hard derivations from the root private key, Master Key

Associated Types

type AccountIndexDerivationType key :: DerivationType

type AddressIndexDerivationType key :: DerivationType

Methods

deriveAccountPrivateKey :: ScrubbedBytes -> key RootK XPrv -> Index (AccountIndexDerivationType key) AccountK -> key AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

deriveAddressPrivateKey :: ScrubbedBytes -> key AccountK XPrv -> AccountingStyle -> Index (AddressIndexDerivationType key) AddressK -> key AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

class HardDerivation key => SoftDerivation (key :: Depth -> * -> *) where

An interface for doing soft derivations from an account public key

Methods

deriveAddressPublicKey :: key AccountK XPub -> AccountingStyle -> Index Soft AddressK -> key AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Helper types

newtype Index (derivationType :: DerivationType) (level :: Depth)

A derivation index, with phantom-types to disambiguate derivation type.

let accountIx = Index 'Hardened 'AccountK
let addressIx = Index 'Soft 'AddressK

Constructors

Index

Fields

getIndex :: Word32

data Depth

Key Depth in the derivation path, according to BIP-0039 / BIP-0044

m | purpose' | cointype' | account' | change | address

We do not manipulate purpose, cointype and change paths directly, so they are left out of the sum type.

Constructors

RootK

AccountK

AddressK

data DerivationType

Type of derivation that should be used with the given indexes.

In theory, we should only consider two derivation types: soft and hard.

However, historically, addresses in Cardano used to be generated across the both soft and hard domain. We therefore introduce a WholeDomain derivation type that is the exact union of Hardened and Soft.

Constructors

Hardened

Soft

WholeDomain

data AccountingStyle

Marker for addresses type engaged. We want to handle three cases here. The first two are pertinent to UTxO accounting and the last one handles rewards from participation in staking. (a) external chain is used for addresses that are part of the advertised targets of a given transaction (b) internal change is for addresses used to handle the change of a the transaction within a given wallet (c) the addresses for a reward (chimeric) account

Constructors

UTxOExternal

UTxOInternal

MutableAccount