Cardano.Address.Derivation

Overview

These abstractions allow generating root private key, also called Master Key and then basing on it enable address derivation

Key Derivation

Types

data Index (derivationType :: DerivationType) (depth :: Depth)

A derivation index, with phantom-types to disambiguate derivation type.

let accountIx = Index 'Hardened 'AccountK
let addressIx = Index 'Soft 'AddressK

Since: 1.0.0

data Depth

Key Depth in the derivation path, according to BIP-0039 / BIP-0044

root | purpose' | cointype' | account' | change | address@
0th      1st         2nd        3rd        4th      5th

We do not manipulate purpose, cointype and change paths directly, so there are no constructors for these.

Since: 1.0.0

Constructors

RootK

AccountK

AddressK

data DerivationType

Type of derivation that should be used with the given indexes.

In theory, we should only consider two derivation types: soft and hard.

However, historically, addresses in Cardano used to be generated across both the soft and the hard domain. We therefore introduce a WholeDomain derivation type that is the exact union of Hardened and Soft.

Since: 1.0.0

Constructors

Hardened

Soft

WholeDomain

data AccountingStyle

Marker for addresses type engaged. We want to handle three cases here. The first two are pertinent to UTxO accounting and the last one handles rewards from participation in staking. (a) external chain is used for addresses that are part of the advertised targets of a given transaction (b) internal change is for addresses used to handle the change of a the transaction within a given wallet (c) the addresses for a reward (chimeric) account

Constructors

UTxOExternal

UTxOInternal

MutableAccount

Abstractions

class GenMasterKey (key :: Depth -> * -> *) where

Abstract interface for constructing a Master Key.

Since: 1.0.0

Associated Types

type SecondFactor key :: *

Methods

genMasterKeyFromMnemonic :: SomeMnemonic -> SecondFactor key -> key RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 1.0.0

genMasterKeyFromXPrv :: XPrv -> key RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 1.0.0

class HardDerivation (key :: Depth -> * -> *) where

An interface for doing hard derivations from the root private key, Master Key

Since: 1.0.0

Associated Types

type AccountIndexDerivationType key :: DerivationType

type AddressIndexDerivationType key :: DerivationType

type WithAccountStyle key :: *

Methods

deriveAccountPrivateKey :: key RootK XPrv -> Index (AccountIndexDerivationType key) AccountK -> key AccountK XPrv

Derives account private key from the given root private key, using derivation scheme 2 (see cardano-crypto package for more details).

Since: 1.0.0

deriveAddressPrivateKey :: key AccountK XPrv -> WithAccountStyle key -> Index (AddressIndexDerivationType key) AddressK -> key AddressK XPrv

Derives address private key from the given account private key, using derivation scheme 2 (see cardano-crypto package for more details).

Since: 1.0.0

class HardDerivation key => SoftDerivation (key :: Depth -> * -> *) where

An interface for doing soft derivations from an account public key

Methods

deriveAddressPublicKey :: key AccountK XPub -> AccountingStyle -> Index Soft AddressK -> key AddressK XPub

Derives address public key from the given account public key, using derivation scheme 2 (see cardano-crypto package for more details).

This is the preferred way of deriving new sequential address public keys.

Since: 1.0.0

Low-Level Cryptography Primitives

XPub

type XPub = XPub

An opaque type representing an extended public key.

Since: 1.0.0

xpubFromBytes :: ByteString -> Maybe XPub

Construct an XPub from raw ByteString

Since: 1.0.0

xpubToBytes :: XPub -> ByteString

Convert an XPub to a raw ByteString

Since: 1.0.0

XPrv

type XPrv = XPrv

An opaque type representing an extended private key.

Since: 1.0.0

xprvFromBytes :: ByteString -> Maybe XPrv

Construct an XPrv from raw ByteString.

Since: 1.0.0

xprvToBytes :: XPrv -> ByteString

Convert an XPrv to a raw ByteString.

Since: 1.0.0

toXPub :: HasCallStack => XPrv -> XPub

Derive the XPub from a corresponding XPrv.

Since: 1.0.0

XSignature

type XSignature = XSignature

An opaque type representing a signature made from an XPrv

Since: 1.0.0

sign :: ByteArrayAccess msg => XPrv -> msg -> XSignature

Produce a signature of the given msg from an XPrv.

Since: 1.0.0

verify :: ByteArrayAccess msg => XPub -> msg -> XSignature -> Bool

Verify the XSignature of a msg with the XPub associated with the XPrv used for signing.

Since: 1.0.0