Cardano.Address.Script

Script

data Script (elem :: Type)

A Script type represents multi signature script. The script embodies conditions that need to be satisfied to make it valid.

Since: 3.0.0

Constructors

RequireSignatureOf !elem

RequireAllOf ![Script elem]

RequireAnyOf ![Script elem]

RequireSomeOf Word8 ![Script elem]

ActiveFromSlot Natural

ActiveUntilSlot Natural

serializeScript :: Script KeyHash -> ByteString

This function realizes what cardano-node's `Api.serialiseToCBOR script` realizes This is basically doing the symbolically following: toCBOR [0,multisigScript]

Since: 3.0.0

foldScript :: (a -> b -> b) -> b -> Script a -> b

Script folding

Since: 3.2.0

Script template

data ScriptTemplate

Represents the script template that show the structure of the script and determines the expected place of verification keys corresponding to given cosigners.

Since: 3.2.0

Constructors

ScriptTemplate

Fields

cosigners :: Map Cosigner XPub

template :: Script Cosigner

newtype Cosigner

Represents the cosigner of the script, ie., party that co-shares the script.

Since: 3.2.0

Constructors

Cosigner Word8

Validation

data ErrValidateScript

Possible validation errors when validating a script

Since: 3.0.0

Constructors

LedgerIncompatible

WrongKeyHash

NotUniformKeyType

Malformed

NotRecommended ErrRecommendedValidateScript

data ErrRecommendedValidateScript

Possible recommended validation errors when validating a script

Since: 3.2.0

Constructors

EmptyList

ListTooSmall

MZero

DuplicateSignatures

RedundantTimelocks

TimelockTrap

data ErrValidateScriptTemplate

Possible validation errors when validating a script template

Since: 3.2.0

Constructors

WrongScript ErrValidateScript

DuplicateXPubs

UnknownCosigner

MissingCosignerXPub

NoCosignerInScript

NoCosignerXPub

validateScript :: ValidationLevel -> Script KeyHash -> Either ErrValidateScript ()

Validate a Script, semantically

Since: 3.0.0

validateScriptTemplate :: ValidationLevel -> ScriptTemplate -> Either ErrValidateScriptTemplate ()

Validate a ScriptTemplate, semantically

Since: 3.2.0

validateScriptOfTemplate :: ValidationLevel -> Script Cosigner -> Either ErrValidateScript ()

Validate a script in ScriptTemplate

Since: 3.5.0

prettyErrValidateScript :: ErrValidateScript -> String

Pretty-print a script validation error.

Since: 3.0.0

prettyErrValidateScriptTemplate :: ErrValidateScriptTemplate -> String

Pretty-print a script template validation error.

Since: 3.2.0

Hashing

newtype ScriptHash

A ScriptHash type represents script hash. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

ScriptHash

Fields

unScriptHash :: ByteString

toScriptHash :: Script KeyHash -> ScriptHash

Computes the hash of a given script, by first serializing it to CBOR.

Since: 3.0.0

scriptHashFromBytes :: ByteString -> Maybe ScriptHash

Construct an ScriptHash from raw ByteString (28 bytes).

Since: 3.0.0

data KeyHash

A KeyHash type represents verification key hash that participate in building multi-signature script. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

KeyHash

Fields

role :: KeyRole

digest :: ByteString

keyHashFromBytes :: (KeyRole, ByteString) -> Maybe KeyHash

Construct an KeyHash from raw ByteString (28 bytes).

Since: 3.0.0

keyHashFromText :: Text -> Either ErrKeyHashFromText KeyHash

Construct a KeyHash from Text. It should be Bech32 encoded text with one of following hrp: - addr_shared_vkh - stake_shared_vkh - addr_vkh - stake_vkh - policy_vkh - addr_shared_vk - stake_shared_vk - addr_vk - stake_vk - addr_shared_xvk - stake_shared_xvk - addr_xvk - stake_xvk - policy_vk - policy_xvk Raw keys will be hashed on the fly, whereas hash that are directly provided will remain as such. If if hex is encountered Unknown policy key is assumed

Since: 3.1.0

keyHashToText :: KeyHash -> Text

Encode a KeyHash to bech32 Text or hex is key role unknown.

Since: 3.0.0