Cardano.Address.Script

Script

data Script

A Script type represents multi signature script. The script embodies conditions that need to be satisfied to make it valid.

Since: 3.0.0

Constructors

RequireSignatureOf !KeyHash

RequireAllOf ![Script]

RequireAnyOf ![Script]

RequireSomeOf Word8 ![Script]

ActiveFromSlot Natural

ActiveUntilSlot Natural

serializeScript :: Script -> ByteString

This function realizes what cardano-node's `Api.serialiseToCBOR script` realizes This is basically doing the symbolically following: toCBOR [0,multisigScript]

Since: 3.0.0

foldScript :: (KeyHash -> b -> b) -> b -> Script -> b

Script folding

Since: 3.2.0

Validation

validateScript :: Script -> Either ErrValidateScript ()

Validate a Script, semantically

Since: 3.0.0

data ErrValidateScript

Possible validation errors when validating a script

Since: 3.0.0

Constructors

EmptyList

ListTooSmall

MZero

DuplicateSignatures

WrongKeyHash

Malformed

prettyErrValidateScript :: ErrValidateScript -> String

Pretty-print a validation error.

Since: 3.0.0

Hashing

newtype ScriptHash

A ScriptHash type represents script hash. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

ScriptHash

Fields

unScriptHash :: ByteString

toScriptHash :: Script -> ScriptHash

Computes the hash of a given script, by first serializing it to CBOR.

Since: 3.0.0

scriptHashFromBytes :: ByteString -> Maybe ScriptHash

Construct an ScriptHash from raw ByteString (28 bytes).

Since: 3.0.0

newtype KeyHash

A KeyHash type represents verification key hash that participate in building multi-signature script. The hash is expected to have size of 28-byte.

Since: 3.0.0

Constructors

KeyHash

Fields

unKeyHash :: ByteString

keyHashFromBytes :: ByteString -> Maybe KeyHash

Construct an KeyHash from raw ByteString (28 bytes).

Since: 3.0.0

keyHashFromText :: Text -> Either ErrKeyHashFromText KeyHash

Construct a KeyHash from Text. Either hex encoded text or Bech32 encoded text with script_vkh, script_vk or script_xvk hrp are expected. Raw keys will be hashed on the fly, whereas hash that are directly provided will remain as such.

Since: 3.1.0

keyHashToText :: KeyHash -> Text

Encode a KeyHash to bech32 Text, using script_vkh as a human readable prefix.

Since: 3.0.0