Cardano.Address.Script.Parser

Script Parser

scriptFromString :: String -> Either ErrValidateScript Script

Run scriptParser on string input.

Since: 3.0.0

scriptParser :: ReadP Script

The script embodies combination of signing keys that need to be met to make it valid. We assume here that the script could delivered from standard input. The examples below are self-explanatory:

requiring signature 3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe

any for signature required any [3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe, 3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3f1]

all signatures required all [3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe, 3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3f1]

at_least 1 signature required at_least 1 [3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe, 3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3f1]

Nested script are supported at_least 1 [3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe, all [3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3f1, 3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3f1]]

Parser is insensitive to whitespaces.

Since: 3.0.0

Script Validator

validateScript :: Script -> Either ErrValidateScript ()

Validate a Script, semantically

Since: 3.0.0

data ErrValidateScript

Possible validation errors when validating a script

Since: 3.0.0

Constructors

EmptyList

ListTooSmall

MZero

DuplicateSignatures

WrongKeyHash

Malformed

prettyErrValidateScript :: ErrValidateScript -> String

Pretty-print a validation error.

Since: 3.0.0