Cardano.Address.Style.Byron

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Byron master keys from mnemonic sentences

HardDerivation: for hierarchical derivation of parent to child keys

PaymentAddress: for constructing addresses from a public key

We call Byron addresses the old address type used by Daedalus in the early days of Cardano. Using this type of addresses and underlying key scheme is now considered deprecated because of some security implications.

Unless you have good reason to do so (like writing backward-compatible code with an existing piece), any new implementation should use the Icarus style for key and addresses.

The internals of the Byron does not matter for the reader, but basically contains what is necessary to perform key derivation and generate addresses from a Byron type.

data Byron (depth :: Depth) key

Deprecated: see Cardano.Address.Style.Icarus

Material for deriving HD random scheme keys, which can be used for making addresses.

Since: 1.0.0

type family DerivationPath (depth :: Depth) :: * where ...

Deprecated: see Cardano.Address.Style.Icarus

The hierarchical derivation indices for a given level/depth.

Since: 1.0.0

Equations

DerivationPath RootK = ()

DerivationPath AccountK = Index WholeDomain AccountK

DerivationPath AddressK = (Index WholeDomain AccountK, Index WholeDomain AddressK)

Accessors

payloadPassphrase :: Byron depth key -> ScrubbedBytes

Deprecated: see Cardano.Address.Style.Icarus

Used for encryption of the derivation path payload within an address.

Since: 1.0.0

derivationPath :: Byron depth key -> DerivationPath depth

Deprecated: see Cardano.Address.Style.Icarus

The address derivation indices for the level of this key.

Since: 1.0.0

getKey :: Byron depth key -> key

Deprecated: see Cardano.Address.Style.Icarus

The raw private or public key.

Since: 1.0.0

Unsafe

liftXPrv :: DerivationPath depth -> XPrv -> Byron depth XPrv

Deprecated: see Cardano.Address.Style.Icarus

Backdoor for generating a new key from a raw XPrv.

Note that the depth is left open so that the caller gets to decide what type of key this is. This is mostly for testing, in practice, seeds are used to represent root keys, and one should genMasterKeyFromXPrv

The first argument is a type-family DerivationPath and its type depends on the depth of the key.

examples:

>>> liftXPrv () prv
_ :: Byron RootK XPrv

>>> liftXPrv minBound prv
_ :: Byron AccountK XPrv

>>> liftXPrv (minBound, minBound) prv
_ :: Byron AddressK XPrv

Since: 1.0.0