Cardano.Address.Style.Icarus

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Icarus master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

PaymentAddress: for constructing addresses from a public key

We call Icarus addresses the new format of Cardano addresses which came after Byron. This is the format initially used in Yoroi and now also used by Daedalus.

Icarus

data Icarus (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Icarus 'RootK XPrv
let accountPubKey  = Icarus 'AccountK XPub
let addressPubKey  = Icarus 'AddressK XPub

Since: 1.0.0

getKey :: Icarus depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 1.0.0

Key Derivation

Generating a root key from SomeMnemonic

:set -XOverloadedStrings
:set -XTypeApplications
:set -XDataKinds
import Cardano.Mnemonic ( mkSomeMnemonic )

let (Right mw) = mkSomeMnemonic @'[15] ["network","empty","cause","mean","expire","private","finger","accident","session","problem","absurd","banner","stage","void","what"]
let sndFactor = mempty -- Or alternatively, a second factor mnemonic transformed to bytes via someMnemonicToBytes
let rootK = genMasterKeyFromMnemonic mw sndFactor :: Icarus 'RootK XPrv

Deriving child keys

Let's consider the following 3rd, 4th and 5th derivation paths 0'/0/14

import Cardano.Address.Derivation ( AccountingStyle(..) )

let accIx = toEnum 0x80000000
let acctK = deriveAccountPrivateKey rootK accIx

let addIx = toEnum 0x00000014
let addrK = deriveAddressPrivateKey acctK UTxOExternal addIx

genMasterKeyFromXPrv :: XPrv -> Icarus RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 1.0.0

genMasterKeyFromMnemonic

Arguments

:: SomeMnemonic
Some valid mnemonic sentence.

-> ScrubbedBytes
An optional second-factor passphrase (or mempty)

-> Icarus RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 1.0.0

deriveAccountPrivateKey :: Icarus RootK XPrv -> Index Hardened AccountK -> Icarus AccountK XPrv

Derives an account private key from the given root private key.

Since: 1.0.0

deriveAddressPrivateKey :: Icarus AccountK XPrv -> AccountingStyle -> Index Soft AddressK -> Icarus AddressK XPrv

Derives an address private key from the given account private key.

Since: 1.0.0

deriveAddressPublicKey :: Icarus AccountK XPub -> AccountingStyle -> Index Soft AddressK -> Icarus AddressK XPub

Derives an address public key from the given account public key.

Since: 1.0.0

Addresses

Generating a PaymentAddress

import Cardano.Address ( bech32 )
import Cardano.Address.Derivation ( AccountingStyle(..), toXPub(..) )

bech32 $ paymentAddress icarusMainnet (toXPub <$> addrK)
"addr1vxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdncxsce5t"

| Analyze an Address to know whether it's an Icarus address or not. Returns Nothing if the address isn't a byron address, or return a structured JSON that gives information about an address.

Since: 2.0.0

paymentAddress :: NetworkDiscriminant Icarus -> Icarus AddressK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Since: 1.0.0

Network Discrimination

icarusMainnet :: NetworkDiscriminant Icarus

NetworkDiscriminant for Cardano MainNet & Icarus

Since: 2.0.0

icarusStaging :: NetworkDiscriminant Icarus

NetworkDiscriminant for Cardano Staging & Icarus

Since: 2.0.0

icarusTestnet :: NetworkDiscriminant Icarus

NetworkDiscriminant for Cardano TestNet & Icarus

Since: 2.0.0

Unsafe

liftXPrv :: XPrv -> Icarus depth XPrv

Unsafe backdoor for constructing an Icarus key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 1.0.0

liftXPub :: XPub -> Icarus depth XPub

Unsafe backdoor for constructing an Icarus key from a raw XPub. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0