Cardano.Address.Style.Shelley

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Shelley master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

paymentAddress: for constructing payment addresses from a address public key or a script

delegationAddress: for constructing delegation addresses from payment credential (public key or script) and stake credential (public key or script)

pointerAddress: for constructing delegation addresses from payment credential (public key or script) and chain pointer

stakeAddress: for constructing reward accounts from stake credential (public key or script)

Shelley

data Shelley (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Shelley 'RootK XPrv
let accountPubKey  = Shelley 'AccountK XPub
let addressPubKey  = Shelley 'PaymentK XPub

Since: 2.0.0

getKey :: Shelley depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 1.0.0

data Role

Describe what the keys within an account are used for.

UTxOExternal: used for public addresses sent to other parties for receiving money.

UTxOInternal: generated by wallet software to send change back to the wallet.

Stake: used for stake key(s) and delegation.

Multisig: used for keys used in multi-party transactions / scripts.

Since: 3.0.0

Constructors

UTxOExternal

UTxOInternal

Stake

Multisig

data family Credential (purpose :: Depth)

Shelley offers several ways to identify ownership of entities on chain.

This data-family has two instances, depending on whether the key is used for payment or for delegation.

Since: 3.0.0

Key Derivation

Generating a root key from SomeMnemonic

:set -XOverloadedStrings
:set -XTypeApplications
:set -XDataKinds
import Cardano.Mnemonic ( mkSomeMnemonic )

let (Right mw) = mkSomeMnemonic @'[15] ["network","empty","cause","mean","expire","private","finger","accident","session","problem","absurd","banner","stage","void","what"]
let sndFactor = mempty -- Or alternatively, a second factor mnemonic transformed to bytes via someMnemonicToBytes
let rootK = genMasterKeyFromMnemonic mw sndFactor :: Shelley 'RootK XPrv

Deriving child keys

Let's consider the following 3rd, 4th and 5th derivation paths 0'/0/14

let accIx = toEnum 0x80000000
let acctK = deriveAccountPrivateKey rootK accIx

let addIx = toEnum 0x00000014
let addrK = deriveAddressPrivateKey acctK UTxOExternal addIx

let stakeK = deriveDelegationPrivateKey acctK

genMasterKeyFromXPrv :: XPrv -> Shelley RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 2.0.0

genMasterKeyFromMnemonic

Arguments

:: SomeMnemonic
Some valid mnemonic sentence.

-> ScrubbedBytes
An optional second-factor passphrase (or mempty)

-> Shelley RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 2.0.0

deriveAccountPrivateKey :: Shelley RootK XPrv -> Index Hardened AccountK -> Shelley AccountK XPrv

Derives an account private key from the given root private key.

Since: 2.0.0

deriveAddressPrivateKey :: Shelley AccountK XPrv -> Role -> Index Soft PaymentK -> Shelley PaymentK XPrv

Derives an address private key from the given account private key.

Since: 2.0.0

deriveDelegationPrivateKey :: Shelley AccountK XPrv -> Shelley DelegationK XPrv

Derive a delegation key for a corresponding AccountK. Note that wallet software are by convention only using one delegation key per account, and always the first account (with index 0').

Deriving delegation keys for something else than the initial account is not recommended and can lead to incompatibility with existing wallet softwares (Daedalus, Yoroi, Adalite...).

Since: 2.0.0

deriveAddressPublicKey :: Shelley AccountK XPub -> Role -> Index Soft PaymentK -> Shelley PaymentK XPub

Derives an address public key from the given account public key.

Since: 2.0.0

deriveMultisigPrivateKey :: Shelley AccountK XPrv -> Index Soft PaymentK -> Shelley ScriptK XPrv

Derives a multisig private key from the given account private key.

Since: 3.0.0

deriveMultisigPublicKey :: Shelley AccountK XPub -> Index Soft PaymentK -> Shelley ScriptK XPub

Derives a multisig public key from the given account public key.

Since: 3.0.0

hashKey :: Shelley key XPub -> KeyHash

Apply Blake2b224 hashing on Shelley XPub. This results in 28-byte hash.

Since: 3.0.0

Addresses

Generating a PaymentAddress from public key credential

import Cardano.Address ( bech32 )
import Cardano.Address.Derivation ( toXPub )

let (Right tag) = mkNetworkDiscriminant 1
let paymentCredential = PaymentFromKey $ (toXPub <$> addrK)
bech32 $ paymentAddress tag paymentCredential
"addr1vxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdncxsce5t"

Generating a PaymentAddress from script credential

import Cardano.ScriptParser ( scriptFromString )
import Cardano.Script ( toScriptHash )
import Codec.Binary.Encoding ( encode, )
import Data.Text.Encoding ( decodeUtf8 )

let (Right tag) = mkNetworkDiscriminant 1
let verKey1 = "3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f3fe"
let verKey2 = "3c07030e36bfffe67e2e2ec09e5293d384637cd2f004356ef320f333"
let scriptStr = "all [" ++ verKey1 ++ ", " ++ verKey2 ++ "]"
let (Just script) = scriptFromString scriptStr
let scriptHash@(ScriptHash bytes) = toScriptHash script
decodeUtf8 (encode EBase16 bytes)
"a015ae61075e25c3d9250bdcbc35c6557272127927ecf2a2d716e29f"
bech32 $ paymentAddress tag (FromScript scriptHash)
"addr1wxspttnpqa0zts7ey59ae0p4ce2hyusj0yn7eu4z6utw98c9uxm83"

Generating a DelegationAddress

let (Right tag) = mkNetworkDiscriminant 1
let paymentCredential = PaymentFromKey $ (toXPub <$> addrK)
let delegationCredential = StakeFromKey $ (toXPub <$> stakeK)
bech32 $ delegationAddress tag paymentCredential delegationCredential
"addr1qxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdn7nudck0fzve4346yytz3wpwv9yhlxt7jwuc7ytwx2vfkyqmkc5xa"

Generating a PointerAddress

import Cardano.Address ( ChainPointer (..) )

let (Right tag) = mkNetworkDiscriminant 1
let ptr = ChainPointer 123 1 2
let paymentCredential = PaymentFromKey $ (toXPub <$> addrK)
bech32 $ pointerAddress tag paymentCredential ptr
"addr1gxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdnmmqypqfcp5um"

inspectShelleyAddress :: (Alternative m, MonadThrow m) => Maybe XPub -> Address -> m Value

Analyze an Address to know whether it's a Shelley address or not.

Throws AddrError if it's not a valid Shelley address, or a ready-to-print string giving details about the Address.

Since: 2.0.0

paymentAddress :: NetworkDiscriminant Shelley -> Credential PaymentK -> Address

Convert a payment credential (key or script) to a payment Address valid for the given network discrimination.

Since: 2.0.0

delegationAddress :: NetworkDiscriminant Shelley -> Credential PaymentK -> Credential DelegationK -> Address

Convert a payment credential (key or script) and a delegation credential (key or script) to a delegation Address valid for the given network discrimination. Funds sent to this address will be delegated according to the delegation settings attached to the delegation key.

Since: 2.0.0

pointerAddress :: NetworkDiscriminant Shelley -> Credential PaymentK -> ChainPointer -> Address

Convert a payment credential (key or script) and pointer to delegation certificate in blockchain to a pointer Address valid for the given network discrimination.

Since: 3.0.0

stakeAddress :: NetworkDiscriminant Shelley -> Credential DelegationK -> Either ErrInvalidStakeAddress Address

Convert a delegation credential (key or script) to a stake Address (aka reward account address) for the given network discrimination.

Since: 3.0.0

extendAddress :: Address -> Credential DelegationK -> Either ErrExtendAddress Address

Extend an existing payment Address to make it a delegation address.

Since: 2.0.0

data ErrExtendAddress

Captures error occuring when trying to extend an invalid address.

Since: 2.0.0

Constructors

ErrInvalidAddressStyle String

ErrInvalidAddressType String

Network Discrimination

newtype MkNetworkDiscriminantError

Error reported from trying to create a network discriminant from number

Since: 2.0.0

Constructors

ErrWrongNetworkTag Integer
Wrong network tag.

mkNetworkDiscriminant :: Integer -> Either MkNetworkDiscriminantError (NetworkDiscriminant Shelley)

Construct NetworkDiscriminant for Cardano Shelley from a number. If the number is invalid, ie., not between 0 and 15, then MkNetworkDiscriminantError is thrown.

Since: 2.0.0

inspectNetworkDiscriminant :: Address -> Maybe (NetworkDiscriminant Shelley)

Retrieve the network discriminant of a given Address. If the Address is malformed or, not a shelley address, returns Nothing.

Since: 2.0.0

shelleyMainnet :: NetworkDiscriminant Shelley

NetworkDicriminant for Cardano MainNet & Shelley

Since: 2.0.0

shelleyTestnet :: NetworkDiscriminant Shelley

NetworkDicriminant for Cardano Testnet & Shelley

Since: 2.0.0

Unsafe

liftXPrv :: XPrv -> Shelley depth XPrv

Unsafe backdoor for constructing an Shelley key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0

liftXPub :: XPub -> Shelley depth XPub

Unsafe backdoor for constructing an Shelley key from a raw XPub. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0