Cardano.Address.Style.Shelley

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Shelley master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

Shelley

data Shelley (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Shelley 'RootK XPrv
let accountPubKey  = Shelley 'AccountK XPub
let addressPubKey  = Shelley 'AddressK XPub

Since: 1.0.0

Accessors

getKey :: Shelley depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 1.0.0

Unsafe

liftXPrv :: XPrv -> Shelley depth XPrv

Unsafe backdoor for constructing an Shelley key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 1.0.0

minSeedLengthBytes :: Int

The minimum seed length for genMasterKeyFromMnemonic.

publicKeySize :: Int

Size, in bytes, of a public key (without chain code)