Cardano.Address.Style.Shelley

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Shelley master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

PaymentAddress: for constructing payment addresses from a address public key

DelegationAddress: for constructing delegation addresses from address and staking public keys

Examples

Generating a root key from SomeMnemonic

:set -XOverloadedStrings
:set -XTypeApplications
:set -XDataKinds
import Cardano.Mnemonic ( mkSomeMnemonic )
import Cardano.Address.Derivation ( GenMasterKey(..) )

let (Right mw) = mkSomeMnemonic @'[15] ["network","empty","cause","mean","expire","private","finger","accident","session","problem","absurd","banner","stage","void","what"]
let sndFactor = mempty -- Or alternatively, a second factor mnemonic transformed to bytes via someMnemonicToBytes
let rootK = genMasterKeyFromMnemonic mw sndFactor :: Shelley 'RootK XPrv

Generating an PaymentAddress and DelegationAddress from a root key

Let's consider the following 3rd, 4th and 5th derivation paths 0'014

import Cardano.Address ( PaymentAddress(..), DelegationAddress (..), bech32)
import Cardano.Address.Derivation ( AccountingStyle(..), StakingDerivation (..), GenMasterKey(..), toXPub )

let accIx = toEnum 0x80000000
let acctK = deriveAccountPrivateKey rootK accIx

let addIx = toEnum 0x00000014
let addrK = deriveAddressPrivateKey acctK UTxOExternal addIx

let (Right netTag) = mkNetworkDiscriminant 1
bech32 $ paymentAddress netTag (toXPub <$> addrK)
"addr1vxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdncxsce5t"

let stakeK = deriveStakingPrivateKey acctK
bech32 $ delegationAddress netTag (toXPub <$> addrK) (toXPub <$> stakeK)
"addr1qxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdn7nudck0fzve4346yytz3wpwv9yhlxt7jwuc7ytwx2vfkyqmkc5xa"

Shelley

data Shelley (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Shelley 'RootK XPrv
let accountPubKey  = Shelley 'AccountK XPub
let addressPubKey  = Shelley 'AddressK XPub

Since: 2.0.0

Discrimination

newtype MkNetworkDiscriminantError

Error reported from trying to create a network discriminant from number

Since: 2.0.0

Constructors

ErrWrongNetworkTag Word8
Wrong network tag.

mkNetworkDiscriminant :: Word8 -> Either MkNetworkDiscriminantError (NetworkDiscriminant Shelley)

Construct NetworkDiscriminant for Cardano Shelley from a number. If the number is invalid, ie., not between 0 and 15, then MkNetworkDiscriminantError is thrown.

Since: 2.0.0

Accessors

getKey :: Shelley depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 1.0.0

Unsafe

liftXPrv :: XPrv -> Shelley depth XPrv

Unsafe backdoor for constructing an Shelley key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0

minSeedLengthBytes :: Int

The minimum seed length for genMasterKeyFromMnemonic.

publicKeyHashSize :: Int

Size, in bytes, of a hash of public key (without the corresponding chain code)