Cardano.Address.Style.Shelley

Description

Documentation

This module provides an implementation of:

GenMasterKey: for generating Shelley master keys from mnemonic sentences

HardDerivation: for hierarchical hard derivation of parent to child keys

SoftDerivation: for hierarchical soft derivation of parent to child keys

PaymentAddress: for constructing payment addresses from a address public key

DelegationAddress: for constructing delegation addresses from address and staking public keys

Shelley

data Shelley (depth :: Depth) key

A cryptographic key for sequential-scheme address derivation, with phantom-types to disambiguate key types.

let rootPrivateKey = Shelley 'RootK XPrv
let accountPubKey  = Shelley 'AccountK XPub
let addressPubKey  = Shelley 'AddressK XPub

Since: 2.0.0

getKey :: Shelley depth key -> key

Extract the raw XPrv or XPub wrapped by this type.

Since: 1.0.0

Key Derivation

Generating a root key from SomeMnemonic

:set -XOverloadedStrings
:set -XTypeApplications
:set -XDataKinds
import Cardano.Mnemonic ( mkSomeMnemonic )

let (Right mw) = mkSomeMnemonic @'[15] ["network","empty","cause","mean","expire","private","finger","accident","session","problem","absurd","banner","stage","void","what"]
let sndFactor = mempty -- Or alternatively, a second factor mnemonic transformed to bytes via someMnemonicToBytes
let rootK = genMasterKeyFromMnemonic mw sndFactor :: Shelley 'RootK XPrv

Deriving child keys

Let's consider the following 3rd, 4th and 5th derivation paths 0'/0/14

import Cardano.Address.Derivation ( AccountingStyle(..) )

let accIx = toEnum 0x80000000
let acctK = deriveAccountPrivateKey rootK accIx

let addIx = toEnum 0x00000014
let addrK = deriveAddressPrivateKey acctK UTxOExternal addIx

let stakeK = deriveStakingPrivateKey acctK

genMasterKeyFromXPrv :: XPrv -> Shelley RootK XPrv

Generate a root key from a corresponding root XPrv

Since: 2.0.0

genMasterKeyFromMnemonic

Arguments

:: SomeMnemonic
Some valid mnemonic sentence.

-> ScrubbedBytes
An optional second-factor passphrase (or mempty)

-> Shelley RootK XPrv

Generate a root key from a corresponding mnemonic.

Since: 2.0.0

deriveAccountPrivateKey :: Shelley RootK XPrv -> Index Hardened AccountK -> Shelley AccountK XPrv

Derives an account private key from the given root private key.

Since: 2.0.0

deriveAddressPrivateKey :: Shelley AccountK XPrv -> AccountingStyle -> Index Soft AddressK -> Shelley AddressK XPrv

Derives an address private key from the given account private key.

Since: 2.0.0

deriveStakingPrivateKey :: Shelley AccountK XPrv -> Shelley StakingK XPrv

Derive a staking key for a corresponding AccountK. Note that wallet software are by convention only using one staking key per account, and always the first account (with index 0').

Deriving staking keys for something else than the initial account is not recommended and can lead to incompatibility with existing wallet softwares (Daedalus, Yoroi, Adalite...).

Since: 2.0.0

deriveAddressPublicKey :: Shelley AccountK XPub -> AccountingStyle -> Index Soft AddressK -> Shelley AddressK XPub

Derives an address public key from the given account public key.

Since: 2.0.0

Addresses

Generating a PaymentAddress

import Cardano.Address ( bech32 )
import Cardano.Address.Derivation ( AccountingStyle(..), toXPub )

let (Right tag) = mkNetworkDiscriminant 1
bech32 $ paymentAddress tag (toXPub <$> addrK)
"addr1vxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdncxsce5t"

Generating a DelegationAddress

import Cardano.Address ( DelegationAddress (..) )
import Cardano.Address.Derivation ( StakingDerivation (..) )

let (Right tag) = mkNetworkDiscriminant 1
bech32 $ delegationAddress tag (toXPub <$> addrK) (toXPub <$> stakeK)
"addr1qxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdn7nudck0fzve4346yytz3wpwv9yhlxt7jwuc7ytwx2vfkyqmkc5xa"

Generating a PointerAddress

import Cardano.Address ( PointerAddress (..), ChainPointer (..) )

let (Right tag) = mkNetworkDiscriminant 1
let ptr = ChainPointer 123 1 2
bech32 $ pointerAddress tag (toXPub <$> addrK) ptr
"addr1gxpfffuj3zkp5g7ct6h4va89caxx9ayq2gvkyfvww48sdnmmqypqfcp5um"

inspectShelleyAddress :: Address -> Maybe Value

Analyze an Address to know whether it's a Shelley address or not.

Returns Nothing if it's not a valid Shelley address, or a ready-to-print string giving details about the Address.

Since: 2.0.0

paymentAddress :: NetworkDiscriminant Shelley -> Shelley AddressK XPub -> Address

Convert a public key to a payment Address valid for the given network discrimination.

Since: 2.0.0

delegationAddress :: NetworkDiscriminant Shelley -> Shelley AddressK XPub -> Shelley StakingK XPub -> Address

Convert a public key and a staking key to a delegation Address valid for the given network discrimination. Funds sent to this address will be delegated according to the delegation settings attached to the delegation key.

Since: 2.0.0

pointerAddress :: NetworkDiscriminant Shelley -> Shelley AddressK XPub -> ChainPointer -> Address

Convert a public key and pointer to staking certificate in blockchain to a pointer Address valid for the given network discrimination.

Since: 2.0.0

extendAddress :: Address -> Either (Shelley StakingK XPub) ChainPointer -> Either ErrExtendAddress Address

Extend an existing payment Address to make it a delegation address.

Since: 2.0.0

data ErrExtendAddress

Captures error occuring when trying to extend an invalid address.

Since: 2.0.0

Constructors

ErrInvalidAddressStyle String

ErrInvalidAddressType String

Network Discrimination

newtype MkNetworkDiscriminantError

Error reported from trying to create a network discriminant from number

Since: 2.0.0

Constructors

ErrWrongNetworkTag Integer
Wrong network tag.

mkNetworkDiscriminant :: Integer -> Either MkNetworkDiscriminantError (NetworkDiscriminant Shelley)

Construct NetworkDiscriminant for Cardano Shelley from a number. If the number is invalid, ie., not between 0 and 15, then MkNetworkDiscriminantError is thrown.

Since: 2.0.0

shelleyMainnet :: NetworkDiscriminant Shelley

NetworkDicriminant for Cardano MainNet & Shelley

Since: 2.0.0

shelleyTestnet :: NetworkDiscriminant Shelley

NetworkDicriminant for Cardano Testnet & Shelley

Since: 2.0.0

Unsafe

liftXPrv :: XPrv -> Shelley depth XPrv

Unsafe backdoor for constructing an Shelley key from a raw XPrv. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0

liftXPub :: XPub -> Shelley depth XPub

Unsafe backdoor for constructing an Shelley key from a raw XPub. this is unsafe because it lets the caller choose the actually derivation depth.

This can be useful however when serializing / deserializing such a type, or to speed up test code (and avoid having to do needless derivations from a master key down to an address key for instance).

Since: 2.0.0